+++ CHANGELOG.md
@@ -11,6 +11,12 @@ This log documents significant changes for each release.  This project follows
 - Issue in interpretation of the path in the resource when a partial resource
   is used as the context, which resulted in incorrect type recognition.
 
+## [4.2.1] - 2024-05-07
+### Fixed
+- The ofType() function now uses automatic conversion when comparing types.
+  See https://hl7.org/fhir/fhirpath.html#types.
+- System.Quantity was treated as a primitive type by mistake.
+
 ## [4.2.0] - 2024-04-30
 ### Added
 - Support for the `LONGNUMBER` literal and the corresponding FHIR `integer64`